move qterm to M-x since I keep forgetting it.
Once upon a time it was C-`. the mtr on the desktop ate that, while
the laptop still used it. M-C-S-1 (aka M-C-!) was the first attempt
at rationalization; had been thinking of figuring out a way to use
M-x for xmonad commands, but I'd want that to include modules so
X.A.Commands doesn't really fly and the whole idea is just a little
too dynamic for xmonad's design. (sigh)

The stuff I'm thinking of for EWMH could also be used to allow
modules to register commands, without forcing a dependency. But
retrofitting that would be a huge job, and there's still things like
getting a list of valid messages and what parameters they take.
Maybe someday.