document plans for sandboxes+XDG+$PATH+xmessage issues
These all interact in such a way that it'd be painful to do them
separately, so get it all over with at once. The new config file
can be used without a functional xmonad.hs, saving us from having
to know how to build xmonad.hs in order to learn how to build
xmonad.hs (which is why this is not in the config record).

New config setup supports full XDG complaince (not hacks; homedir
can be shared) and also fits into OS X with some path adjustments.

Design still a work in progress. No code since design isn't final.