still tweaking display stuff
And it turns out I've had two monitors in a row
that _don't support DPMI properly_:
DPMI them suspend or deeper, and they disconnect
and everything else gets turned back on by rescreen,
which then turns the buggy monitor back on
again. No idea how to deal with this as yet.

On top of which, it _moves around_ in `xrandr` so now
I need to do a lot of extra work in arandr setup.